refactor(badger): split repositories into separate files
